Create E2E LoRa network data from within tests.
