+++ test2.0/e2e/specs/001-multi-networks.js
@@ -2,10 +2,12 @@
 var assert = require('assert')
 var chai = require('chai')
 var chaiHttp = require('chai-http')
-var app = require('../../restApp.js')
+var app = require('../../../restApp.js')
 var should = chai.should()
-var setup = require('./setup.js')
-var appLogger = require('../../rest/lib/appLogger.js')
+var setup = require('../setup.js')
+var appLogger = require('../../../rest/lib/appLogger.js')
+const Lora1 = require('../networks/lora-v1')
+const Lora2 = require('../networks/lora-v2')
 
 chai.use(chaiHttp)
 var server = chai.request(app).keepOpen()
@@ -57,64 +59,51 @@ describe('E2E Test for Multiple Networks', () => {
   before(async () => {
     await setup.start()
     await wait(10000)
-
+    await Lora1.setup()
+    await Lora2.setup()
   })
+
   describe('Verify Login and Administration of Users Works', () => {
-    it('Admin Login to LPWan Server', (done) => {
-      server
+    it('Admin Login to LPWan Server', async () => {
+      const res = await server
         .post('/api/sessions')
         .send({'login_username': 'admin', 'login_password': 'password'})
-        .end(function (err, res) {
-          if (err) done(err)
-          res.should.have.status(200)
-          adminToken = res.text
-          done()
-        })
+      res.should.have.status(200)
+      adminToken = res.text
     })
 
-    it('Create a Application User Account', (done) => {
-      server
+    it('Create a Application User Account', async () => {
+      const res = await server
         .post('/api/users')
         .set('Authorization', 'Bearer ' + adminToken)
         .set('Content-Type', 'application/json')
         .send({ 'username': 'bobmouse', 'password': 'mousetrap', 'role': 'user', 'companyId': 1 })
-        .end(function (err, res) {
-          if (err) done(err)
-          res.should.have.status(200)
-          let ret = JSON.parse(res.text)
-          ret.should.have.property('id')
-          userId = ret.id
-          done()
-        })
+      res.should.have.status(200)
+      let ret = JSON.parse(res.text)
+      ret.should.have.property('id')
+      userId = ret.id
     })
-    it('Verify Application User Exists', (done) => {
-      server
+    it('Verify Application User Exists', async () => {
+      const res = await server
         .get('/api/users/' + userId)
         .set('Authorization', 'Bearer ' + adminToken)
         .set('Content-Type', 'application/json')
         .send()
-        .end(function (err, res) {
-          if (err) done(err)
-          res.should.have.status(200)
-          let userObj = JSON.parse(res.text)
-          userObj.username.should.equal('bobmouse')
-          userObj.role.should.equal('user')
-          done()
-        })
+      res.should.have.status(200)
+      let userObj = JSON.parse(res.text)
+      userObj.username.should.equal('bobmouse')
+      userObj.role.should.equal('user')
     })
-    it('Application User Login to LPWan Server', (done) => {
-      server
+    it('Application User Login to LPWan Server', async () => {
+      const res = await server
         .post('/api/sessions')
         .send({'login_username': 'bobmouse', 'login_password': 'mousetrap'})
-        .end(function (err, res) {
-          if (err) done(err)
-          res.should.have.status(200)
-          userToken = res.text
-          should.exist(userToken)
-          done()
-        })
+      res.should.have.status(200)
+      userToken = res.text
+      should.exist(userToken)
     })
   })
+
   describe('Setup Networks', () => {
     describe('Setup Lora 1.0 Network', () => {
       it('Create Network Provider', async () => {
@@ -125,24 +114,20 @@ describe('E2E Test for Multiple Networks', () => {
           .send({ 'name': 'Kyrio' })
         netProvId = JSON.parse(res.text).id
       })
-      it('Verify LoraOS 1.0 Protocol Exists', (done) => {
-        server
+      it('Verify LoraOS 1.0 Protocol Exists', async () => {
+        const res = await server
           .get('/api/networkProtocols?search=LoRa Server&networkProtocolVersion=1.0')
           .set('Authorization', 'Bearer ' + adminToken)
           .set('Content-Type', 'application/json')
-          .end(function (err, res) {
-            if (err) done(err)
-            res.should.have.status(200)
-            let result = JSON.parse(res.text)
-            appLogger.log(result)
-            result.records.should.be.instanceof(Array)
-            result.records.should.have.length(1)
-            result.totalCount.should.equal(1)
-            result.records[0].should.have.property('networkProtocolVersion')
-            result.records[0].networkProtocolVersion.should.equal('1.0')
-            lora.loraV1.protocolId = result.records[0].id
-            done()
-          })
+        res.should.have.status(200)
+        let result = JSON.parse(res.text)
+        appLogger.log(result)
+        result.records.should.be.instanceof(Array)
+        result.records.should.have.length(1)
+        result.totalCount.should.equal(1)
+        result.records[0].should.have.property('networkProtocolVersion')
+        result.records[0].networkProtocolVersion.should.equal('1.0')
+        lora.loraV1.protocolId = result.records[0].id
       })
       it('Create the Local LoraOS 1.0 Network', (done) => {
         server
@@ -153,9 +138,9 @@ describe('E2E Test for Multiple Networks', () => {
             'name': 'LocalLoraOS1_0',
             'networkProviderId': netProvId,
             'networkTypeId': 1,
-            'baseUrl': 'https://lora_appserver1:8080/api',
+            'baseUrl': Lora1.network.baseUrl,
             'networkProtocolId': lora.loraV1.protocolId,
-            'securityData': {authorized: false, 'username': 'admin', 'password': 'admin'}
+            'securityData': { authorized: false, ...Lora1.account }
           })
           .end(function (err, res) {
             if (err) done(err)
@@ -180,7 +165,7 @@ describe('E2E Test for Multiple Networks', () => {
             res.should.have.status(200)
             let network = JSON.parse(res.text)
             network.name.should.equal('LocalLoraOS1_0')
-            network.baseUrl.should.equal('https://lora_appserver1:8080/api')
+            network.baseUrl.should.equal(Lora1.network.baseUrl)
             network.securityData.authorized.should.equal(true)
             network.securityData.message.should.equal('ok')
             network.securityData.enabled.should.equal(true)
@@ -217,9 +202,9 @@ describe('E2E Test for Multiple Networks', () => {
             'name': 'LocalLoraOS2_0',
             'networkProviderId': netProvId,
             'networkTypeId': 1,
-            'baseUrl': 'https://lora_appserver:8080/api',
+            'baseUrl': Lora2.network.baseUrl,
             'networkProtocolId': lora.loraV2.protocolId,
-            'securityData': {authorized: false, 'username': 'admin', 'password': 'admin'}
+            'securityData': { authorized: false, ...Lora2.account }
           })
           .end(function (err, res) {
             if (err) done(err)
@@ -244,7 +229,7 @@ describe('E2E Test for Multiple Networks', () => {
             res.should.have.status(200)
             let network = JSON.parse(res.text)
             network.name.should.equal('LocalLoraOS2_0')
-            network.baseUrl.should.equal('https://lora_appserver:8080/api')
+            network.baseUrl.should.equal(Lora2.network.baseUrl)
             network.securityData.authorized.should.equal(true)
             network.securityData.message.should.equal('ok')
             network.securityData.enabled.should.equal(true)
@@ -447,15 +432,9 @@ describe('E2E Test for Multiple Networks', () => {
             applications.should.have.property('totalCount')
             applications.should.have.property('records')
             // applications.totalCount.should.equal(2)
-            let application = {}
-            for (let index = 0; index < applications.records.length; index++) {
-              if (applications.records[index].name === 'BobMouseTrapLv1') {
-                application = applications.records[index]
-              }
-            }
-            application.should.have.property('name')
-            application.name.should.equal('BobMouseTrapLv1')
-            application.description.should.equal('CableLabs Test Application')
+            let application = applications.records.find(x => x.name === Lora1.application.name)
+            application.name.should.equal(Lora1.application.name)
+            application.description.should.equal(Lora1.application.description)
             lora.loraV1.apps.push({
               appId: application.id,
               appNTLId: '',
@@ -507,15 +486,8 @@ describe('E2E Test for Multiple Networks', () => {
             appNTLs.should.have.property('totalCount')
             appNTLs.should.have.property('records')
             // appNTLs.totalCount.should.equal(2)
-            let appNTL = {}
-            for (let index = 0; index < appNTLs.records.length; index++) {
-              if (appNTLs.records[index].applicationId === lora.loraV1.apps[0].appId) {
-                appNTL = appNTLs.records[index]
-              }
-            }
+            let appNTL = appNTLs.records.find(x => x.applicationId === lora.loraV1.apps[0].appId)
             should.exist(appNTL)
-            appLogger.log(appNTL)
-            // appNTL.should.eql(expected)
             lora.loraV1.apps[0].appNTLId = appNTL.id
             done()
           })
@@ -536,16 +508,10 @@ describe('E2E Test for Multiple Networks', () => {
             applications.should.have.property('records')
             appLogger.log(applications, 'error')
             // applications.totalCount.should.equal(2)
-            let application = {}
-            for (let index = 0; index < applications.records.length; index++) {
-              if (applications.records[index].name === 'BobMouseTrapLv2') {
-                application = applications.records[index]
-              }
-            }
+            let application = applications.records.find(x => x.name === Lora2.application.name)
             should.exist(application)
-            appLogger.log(application)
-            application.name.should.equal('BobMouseTrapLv2')
-            application.description.should.equal('CableLabs Test Application')
+            application.name.should.equal(Lora2.application.name)
+            application.description.should.equal(Lora2.application.description)
             lora.loraV2.apps.push({
               appId: application.id,
               appNTLId: '',
@@ -597,15 +563,8 @@ describe('E2E Test for Multiple Networks', () => {
             appNTLs.should.have.property('totalCount')
             appNTLs.should.have.property('records')
             // appNTLs.totalCount.should.equal(2)
-            let appNTL = {}
-            for (let index = 0; index < appNTLs.records.length; index++) {
-              if (appNTLs.records[index].applicationId === lora.loraV2.apps[0].appId) {
-                appNTL = appNTLs.records[index]
-              }
-            }
+            let appNTL = appNTLs.records.find(x => x.applicationId === lora.loraV2.apps[0].appId)
             should.exist(appNTL)
-            appLogger.log(appNTL)
-            // appNTL.should.eql(expected)
             lora.loraV2.apps[0].appNTLId = appNTL.id
             done()
           })
@@ -788,19 +747,14 @@ describe('E2E Test for Multiple Networks', () => {
             deviceProfiles.should.have.property('totalCount')
             deviceProfiles.should.have.property('records')
             // deviceProfiles.totalCount.should.equal(2)
-            let deviceProfile = {}
-            for (let index = 0; index < deviceProfiles.records.length; index++) {
-              if (deviceProfiles.records[index].name === 'BobMouseTrapDeviceProfileLv1') {
-                deviceProfile = deviceProfiles.records[index]
-              }
-            }
+            let deviceProfile = deviceProfiles.records.find(x => x.name === Lora1.deviceProfile.name)
             should.exist(deviceProfile)
             deviceProfile.description.should.equal('Device Profile managed by LPWAN Server, perform changes via LPWAN')
             deviceProfile.id.should.equal(1)
-            deviceProfile.name.should.equal('BobMouseTrapDeviceProfileLv1')
-            deviceProfile.networkSettings.name.should.equal('BobMouseTrapDeviceProfileLv1')
-            deviceProfile.networkSettings.networkServerID.should.equal('5')
-            deviceProfile.networkSettings.organizationID.should.equal('56')
+            deviceProfile.name.should.equal(Lora1.deviceProfile.name)
+            deviceProfile.networkSettings.name.should.equal(Lora1.deviceProfile.name)
+            deviceProfile.networkSettings.networkServerID.should.equal(Lora1.deviceProfile.networkServerID)
+            deviceProfile.networkSettings.organizationID.should.equal(Lora1.deviceProfile.organizationID)
             lora.loraV1.apps[0].deviceProfileIds.push(deviceProfile.id)
             done()
           })
@@ -809,9 +763,9 @@ describe('E2E Test for Multiple Networks', () => {
         let expected = {
           'id': 1,
           'applicationId': lora.loraV1.apps[0].appId,
-          'name': 'BobMouseTrapDeviceLv1',
+          'name': Lora1.device.name,
           'deviceModel': null,
-          'description': 'Test Device for E2E'
+          'description': Lora1.device.description
         }
         server
           .get('/api/devices')
@@ -825,14 +779,8 @@ describe('E2E Test for Multiple Networks', () => {
             devices.should.have.property('totalCount')
             devices.should.have.property('records')
             // devices.totalCount.should.equal(2)
-            let device = {}
-            for (let index = 0; index < devices.records.length; index++) {
-              if (devices.records[index].name === 'BobMouseTrapDeviceLv1') {
-                device = devices.records[index]
-              }
-            }
+            let device = devices.records.find(x => x.name === Lora1.device.name)
             should.exist(device)
-            appLogger.log(device)
             device.should.eql(expected)
             lora.loraV1.apps[0].deviceIds.push(device.id)
             done()
@@ -878,15 +826,8 @@ describe('E2E Test for Multiple Networks', () => {
             deviceNTLs.should.have.property('totalCount')
             deviceNTLs.should.have.property('records')
             // deviceNTLs.totalCount.should.equal(2)
-            let deviceNTL = {}
-            for (let index = 0; index < deviceNTLs.records.length; index++) {
-              if (deviceNTLs.records[index].deviceId === lora.loraV1.apps[0].deviceIds[0]) {
-                deviceNTL = deviceNTLs.records[index]
-              }
-            }
+            let deviceNTL = deviceNTLs.records.find(x => x.deviceId === lora.loraV1.apps[0].deviceIds[0])
             should.exist(deviceNTL)
-            appLogger.log(deviceNTL)
-            // deviceNTL.should.eql(expected)
             lora.loraV1.apps[0].deviceNTLIds.push(deviceNTL.id)
             done()
           })
@@ -930,19 +871,14 @@ describe('E2E Test for Multiple Networks', () => {
           .set('Authorization', 'Bearer ' + adminToken)
           .set('Content-Type', 'application/json')
           .end(function (err, res) {
-            if (err) done(err)
+            if (err) return done(err)
             res.should.have.status(200)
             res.should.have.property('text')
             let deviceProfiles = JSON.parse(res.text)
             deviceProfiles.should.have.property('totalCount')
             deviceProfiles.should.have.property('records')
             // deviceProfiles.totalCount.should.equal(2)
-            let deviceProfile = {}
-            for (let index = 0; index < deviceProfiles.records.length; index++) {
-              if (deviceProfiles.records[index].name === 'BobMouseTrapDeviceProfileLv2') {
-                deviceProfile = deviceProfiles.records[index]
-              }
-            }
+            let deviceProfile = deviceProfiles.records.find(x => x.name === Lora2.deviceProfile.name)
             should.exist(deviceProfile)
             // deviceProfile.should.eql(expected)
             lora.loraV2.apps[0].deviceProfileIds.push(deviceProfile.id)
@@ -969,12 +905,7 @@ describe('E2E Test for Multiple Networks', () => {
             devices.should.have.property('totalCount')
             devices.should.have.property('records')
             // devices.totalCount.should.equal(2)
-            let device = {}
-            for (let index = 0; index < devices.records.length; index++) {
-              if (devices.records[index].name === 'BobMouseTrapDeviceLv2') {
-                device = devices.records[index]
-              }
-            }
+            let device = devices.records.find(x => x.name === Lora2.device.name)
             should.exist(device)
             appLogger.log(device)
             // device.should.eql(expected)
@@ -1023,12 +954,7 @@ describe('E2E Test for Multiple Networks', () => {
             deviceNTLs.should.have.property('totalCount')
             deviceNTLs.should.have.property('records')
             // deviceNTLs.totalCount.should.equal(2)
-            let deviceNTL = {}
-            for (let index = 0; index < deviceNTLs.records.length; index++) {
-              if (deviceNTLs.records[index].deviceId === lora.loraV2.apps[0].deviceIds[0]) {
-                deviceNTL = deviceNTLs.records[index]
-              }
-            }
+            let deviceNTL = deviceNTLs.records.find(x => x.deviceId === lora.loraV2.apps[0].deviceIds[0])
             should.exist(deviceNTL)
             appLogger.log(deviceNTL)
             // deviceNTL.should.eql(expected)
@@ -1212,16 +1138,8 @@ describe('E2E Test for Multiple Networks', () => {
             devices.should.have.property('records')
             appLogger.log(devices)
             // devices.totalCount.should.equal(2)
-            let device1 = {}
-            let device2 = {}
-            for (let index = 0; index < devices.records.length; index++) {
-              if (devices.records[index].name === '00B7641AD008A5FC') {
-                device1 = devices.records[index]
-              }
-              else if (devices.records[index].name === '1234567890987654') {
-                device2 = devices.records[index]
-              }
-            }
+            let device1 = devices.records.find(x => x.name === '00B7641AD008A5FC')
+            let device2 = devices.records.find(x => x.name === '1234567890987654')
             should.exist(device1)
             should.exist(device2)
             lora.ttn.apps[0].deviceIds.push(device1.id)
@@ -1273,12 +1191,7 @@ describe('E2E Test for Multiple Networks', () => {
             appLogger.log(deviceNTLs)
 
             // deviceNTLs.totalCount.should.equal(2)
-            let deviceNTL = {}
-            for (let index = 0; index < deviceNTLs.records.length; index++) {
-              if (deviceNTLs.records[index].deviceId === lora.loraV2.apps[0].deviceIds[0]) {
-                deviceNTL = deviceNTLs.records[index]
-              }
-            }
+            let deviceNTL = deviceNTLs.records.find(x => x.deviceId === lora.loraV2.apps[0].deviceIds[0])
             should.exist(deviceNTL)
             appLogger.log(deviceNTL)
             // deviceNTL.should.eql(expected)
